% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/kinship_functions.R
\name{coeffInbr}
\alias{coeffInbr}
\title{CoeffInbr function}
\usage{
coeffInbr(pedigree)
}
\arguments{
\item{pedigree}{}
}
\description{
Generates a vector of length n of inbreeding coefficients for n individuals of a pedigree
}
\examples{
coeffInbr()
}

