% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/kinship_functions.R
\name{kinship}
\alias{kinship}
\title{Kinship function}
\usage{
kinship(pedigree)
}
\arguments{
\item{pedigree}{}
}
\description{
Generates a n*n matrix of kinship coefficients for n individuals of a pedigree
}
\examples{
kinship()
}

