\name{pedigree-package}
\alias{pedigree-package}
\alias{pedigree}
\docType{package}
\title{
  Package to deal with pedigree data}
\description{
  Package with functions to analyse and transform pedigree data. A
  pedigree is a \code{data.frame } where the first column contains an
  ID, and the second and third columns contain ID of first and second parent.
}

\author{
  Albart Coster: <albart.coster@wur.nl>
}
\keyword{utilities}
\seealso{
  \code{\link[pedigree]{trimPed}}
  \code{\link[pedigree]{orderPed}}
  \code{\link[pedigree]{countGen}}
  \code{\link[pedigree]{makeA}}
  \code{\link[pedigree]{makeAinv}}
  \code{\link[pedigree]{calcInbreeding}}
  \code{\link[pedigree]{add.Inds}}
}

