% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/kinship_functions.R
\name{scanped}
\alias{scanped}
\title{Scanped Function}
\usage{
scanped(filename)
}
\arguments{
\item{filename}{}

\item{header.}{Defaults to TRUE}
}
\description{
Reads in a pedigree from a specified input file.
If parameter header=TRUE the first line should contain a header line of the form:
Indiv  Father  Mother  Sex
Each subsequent line should contain 3 integer indexes (individual, father, mother) and a letter M or F (sex)
All individuals must have an integer label larger than that of any ancestor and the k founders should have labels 1..k
}
\examples{
scanped()
}

